+++ packages/common/src/index.ts
@@ -3,10 +3,6 @@ export * from './blockstore/memory-blockstore'
 export * from './blockstore/persistent-blockstore'
 
 export * from './repo'
-export * from './repo/tid'
-export * from './repo/mst'
-export * from './repo/types'
-export * as delta from './repo/delta'
 
 export * from './microblog/index'
 export * from './microblog/types'
